Assignments nearly done, need more implementation from courses

+++ api/assignments.js
@@ -5,23 +5,21 @@ const { validateAgainstSchema, extractValidFields } = require('../lib/validation
 const { requireAuthentication } = require('../lib/auth')
 const {
   AssignmentSchema,
+  EditableAssignmentSchema,
   insertNewAssignment,
+  insertNewSubmission,
   getAssignmentById,
+  getAllSubmissions,
   updateAssignmentById,
   deleteAssignmentById
 } = require('../models/assignment')
 
 const {
-    UserSchema, 
-    insertNewUser,
-    getUserById, 
-    validateUser, 
-    getStudentCourses, 
-    getInstructorCourses, 
-    getUserByEmail
+    getUserById
 } = require('../models/user')
 
 const { getCourseById } = require('../models/course')
+const req = require('express/lib/request')
 
 const router = Router()
 
@@ -41,6 +39,16 @@ async function isUserAuthorized(userId, courseId){
             return false    
     }
 } 
+
+async function isStudent(userId, courseId){
+    const reqUser = await getUserById(userId, false)
+    const course = await getCourseById(courseId)
+    if(reqUser.role == "student"){
+        //determine if student is in the course
+        return true
+    }
+    return false
+}
 
 router.post('/', requireAuthentication, async function (req, res, next) {
   const isAuthorized = await isUserAuthorized(req.user, req.body.courseId)
@@ -64,9 +72,11 @@ router.get('/:id', async function (req, res, next) {
   }
 })
 
-router.patch('/:id', async function (req, res, next) {
+router.patch('/:id', requireAuthentication, async function (req, res, next) {
   const id = req.params.id
-  if (validateAgainstSchema(req.body, AssignmentSchema)) {
+  const isAuthorized = await isUserAuthorized(req.user, req.body.courseId)
+  // Cannot Edit Submissions within this endpoint
+  if (validateAgainstSchema(req.body, EditableAssignmentSchema) && isAuthorized) {
     await updateAssignmentById(id, req.body)
     res.status(200).send({ AssignmentUpdated: "Assignment Sucessfully Updated!"})
 } else {
@@ -77,10 +87,37 @@ router.patch('/:id', async function (req, res, next) {
 
 })
 
-router.delete('/:id', async function (req, res, next) {
+router.delete('/:id', requireAuthentication, async function (req, res, next) {
   const id = req.params.id
-  await deleteAssignmentById(id)
+  const assignment = await getAssignmentById(id)
+  const isAuthorized = await isUserAuthorized(req.user, assignment.courseId)
+  if(isAuthorized){
+    await deleteAssignmentById(id)
+  }
   res.status(204).send()
+})
+
+router.get('/:id/submissions', requireAuthentication, async function (req, res, next) {
+    const id = req.params.id
+    const isAuthorized = await isUserAuthorized(req.user, assignment.courseId)
+    const submissions = await getAllSubmissions(id, req.query.page)
+    if (submissions && isAuthorized) {
+        res.status(200).send({submissions: submissions})
+    } else {
+        next()
+    }
+})
+
+router.post('/:id/submissions', requireAuthentication, async function (req, res, next) {
+  const isAuthorized = await isStudent(req.user, req.body.courseId)
+  if (validateAgainstSchema(req.body, AssignmentSchema) && isAuthorized) {
+      const id = await insertNewSubmission(req.body)
+      res.status(201).send({ id: id })
+  } else {
+      res.status(400).send({
+          err: "Request body is not a valid Submission"
+      })
+  }
 })
 
 module.exports = router
